+++ lib/test_maple_tree.c
@@ -1574,7 +1574,7 @@ static noinline void __init check_root_expand(struct maple_tree *mt)
 	mas_set(&mas, 1);
 	ptr = mas_prev(&mas, 0);
 	MT_BUG_ON(mt, (mas.index != 0) && (mas.last != 0));
-	MT_BUG_ON(mt, ptr != (void *)((unsigned long) check_prev_entry | 1UL));
+	MT_BUG_ON(mt, ptr != (void *)(uintptr_t)((unsigned long) check_prev_entry | 1UL));
 
 	mas_unlock(&mas);
 
@@ -1592,7 +1592,7 @@ static noinline void __init check_root_expand(struct maple_tree *mt)
 	mas_set(&mas, 1);
 	ptr = mas_prev(&mas, 0);
 	MT_BUG_ON(mt, (mas.index != 0) && (mas.last != 0));
-	MT_BUG_ON(mt, ptr != (void *)((unsigned long) check_prev_entry | 2UL));
+	MT_BUG_ON(mt, ptr != (void *)(uintptr_t)((unsigned long) check_prev_entry | 2UL));
 
 
 	mas_unlock(&mas);
